moved GetSplits helpers into helpers.go
